+++ Games/EchoGame/src/main/java/jokatu/game/games/echo/Echo.java
@@ -2,6 +2,7 @@ package jokatu.game.games.echo;
 
 import jokatu.game.event.PublicGameEvent;
 import jokatu.game.games.echo.input.EchoInput;
+import jokatu.game.games.echo.player.EchoPlayer;
 import org.jetbrains.annotations.NotNull;
 
 public class Echo implements PublicGameEvent {
