+++ Games/EchoGame/src/main/java/jokatu/game/games/echo/input/EchoInputAcceptor.java
@@ -2,7 +2,7 @@ package jokatu.game.games.echo.input;
 
 import jokatu.game.event.AbstractPrivateGameEvent;
 import jokatu.game.games.echo.Echo;
-import jokatu.game.games.echo.EchoPlayer;
+import jokatu.game.games.echo.player.EchoPlayer;
 import jokatu.game.input.InputAcceptor;
 import org.jetbrains.annotations.NotNull;
