+++ Games/EchoGame/src/main/java/jokatu/game/games/echo/player/EchoPlayer.java
@@ -1,4 +1,4 @@
-package jokatu.game.games.echo;
+package jokatu.game.games.echo.player;
 
 import jokatu.game.player.AbstractPlayer;
