+++ Games/EchoGame/src/main/java/jokatu/game/games/echo/views/EchoViewResolverFactory.java
@@ -1,7 +1,7 @@
 package jokatu.game.games.echo.views;
 
 import jokatu.game.games.echo.EchoGame;
-import jokatu.game.games.echo.EchoPlayer;
+import jokatu.game.games.echo.player.EchoPlayer;
 import jokatu.game.viewresolver.ViewResolver;
 import jokatu.game.viewresolver.ViewResolverFactory;
 import org.jetbrains.annotations.NotNull;
